fix: 限界省メモリ化

+++ program-a/src/geotag.rs
@@ -6,8 +6,8 @@ use std::cmp::Ordering;
 pub struct Geotag {
     pub id: u64,
     pub date: String,
-    pub latitude: f64,
-    pub longitude: f64,
+    pub latitude: f32,
+    pub longitude: f32,
     pub url: String,
 }
